Implements product filtering, dynamic tab icon cls, and mock inventory data
The grid is temporary and the filtering doesn't seem to work on any tab except the mugs tab, likely a sync problem. Not gonna bother fixing it because the grid is a placeholder. Next push will contain individual item "cards" on the main screen with a product picture, description, price, and buy button.